+++ gluuengine/dockerclient/_docker.py
@@ -4,14 +4,18 @@
 # All rights reserved.
 
 import json
+import os
+import shutil
+import tempfile
 from collections import namedtuple
 from contextlib import contextmanager
 
-from docker import Client
+import docker
 
 from ..errors import DockerExecError
 from ..registry import REGISTRY_BASE_URL
-from ..utils import po_run
+from ..utils import make_tarfile
+from ..utils import extract_tarfile
 
 
 DockerExecResult = namedtuple("DockerExecResult",
@@ -55,15 +59,15 @@ class Docker(object):
             hostname=hostname,
         )
 
-    # def get_container_ip(self, container_id):
-    #     """Gets container IP.
+    def get_container_ip(self, container_id):
+        """Gets container IP.
 
-    #     :param container_id: ID or name of the container.
-    #     :returns: Container's IP address.
-    #     """
-    #     with self._get_client() as client:
-    #         info = client.inspect_container(container_id)
-    #         return info["NetworkSettings"]["IPAddress"]
+        :param container_id: ID or name of the container.
+        :returns: Container's IP address.
+        """
+        with self._get_client() as client:
+            info = client.inspect_container(container_id)
+            return info["NetworkSettings"]["Networks"]["weave"]["IPAddress"]
 
     def remove_container(self, container_id):
         """Removes container.
@@ -158,16 +162,39 @@ class Docker(object):
             return container_id
 
     def copy_to_container(self, container, src, dest):
-        cfg_str = self._swarm_conf_str()
-        cmd = "docker {} cp {} {}:{}".format(cfg_str, src, container, dest)
-        stdout, stderr, err_code = po_run(cmd)
-        return stdout, stderr, err_code
+        res = self.exec_cmd(container, "mktemp -d")
+        tmp_path = res.retval
+
+        with make_tarfile(src) as tf:
+            with self._get_client() as client:
+                client.put_archive(container, tmp_path, tf)
+
+        self.exec_cmd(
+            container,
+            "mkdir -p {}".format(os.path.dirname(dest)),
+        )
+        self.exec_cmd(
+            container,
+            "mv {}/{} {}".format(tmp_path, os.path.basename(src), dest),
+        )
+        self.exec_cmd(container, "rm -rf {}".format(tmp_path))
 
     def copy_from_container(self, container, src, dest):
-        cfg_str = self._swarm_conf_str()
-        cmd = "docker {} cp {}:{} {}".format(cfg_str, container, src, dest)
-        stdout, stderr, err_code = po_run(cmd)
-        return stdout, stderr, err_code
+        with tempfile.NamedTemporaryFile() as fd:
+            with self._get_client() as client:
+                resp, _ = client.get_archive(container, src)
+                for stream in resp:
+                    fd.write(stream)
+                fd.seek(0)
+
+                # pull archive to temporary path
+                tmp_path = tempfile.mkdtemp()
+                extract_tarfile(fd, tmp_path)
+
+                if not os.path.exists(os.path.dirname(dest)):
+                    os.makedirs(os.path.dirname(dest))
+                shutil.move("{}/{}".format(tmp_path, os.path.basename(src)), dest)
+                shutil.rmtree(tmp_path)
 
     def _swarm_conf_str(self):
         cfg_str = " ".join([
@@ -203,7 +230,7 @@ class Docker(object):
         else:
             cfg = self.config
 
-        client = Client(base_url=cfg.get("base_url"), tls=cfg.get("tls"))
+        client = docker.Client(base_url=cfg.get("base_url"), tls=cfg.get("tls"))
         try:
             yield client
         finally: